requirements.txt
Django>=2.0
django-allauth==0.35
django-bootstrap3==9.1
django-crispy-forms
django-formtools
Pillow
psycopg2